options in autoComplete based on const
