+++ designer/client/src/components/toolbars/search/AdvancedSearchFilters.tsx
@@ -1,11 +1,16 @@
 import React, { useEffect, useMemo } from "react";
-import { Button, Box, Typography } from "@mui/material";
+import { Box, Button, Typography } from "@mui/material";
 import { useTranslation } from "react-i18next";
 import { SearchLabeledInput } from "../../sidePanels/SearchLabeledInput";
 import { SearchLabel } from "../../sidePanels/SearchLabel";
-import { resolveSearchQuery, searchQueryToString, useNodeTypes } from "./utils";
+import { resolveSearchQuery, searchQueryToString, selectorByName } from "./utils";
 import { SearchQuery } from "./SearchResults";
 import { SearchLabeledAutocomplete } from "../../sidePanels/SearchLabeledAutocomplete";
+import { useSelector } from "react-redux";
+import { getScenario } from "../../../reducers/selectors/graph";
+import NodeUtils from "../../graph/NodeUtils";
+import { uniq } from "lodash";
+import { getComponentGroups } from "../../../reducers/selectors/settings";
 
 export function AdvancedSearchFilters({
     filterFields,
@@ -21,6 +26,9 @@ export function AdvancedSearchFilters({
     setCollapsedHandler: React.Dispatch<React.SetStateAction<boolean>>;
 }) {
     const { t } = useTranslation();
+    const componentsGroups = useSelector(getComponentGroups);
+    const { scenarioGraph } = useSelector(getScenario);
+    const allNodes = NodeUtils.nodesFromScenarioGraph(scenarioGraph);
 
     const displayNames = useMemo(
         () => ({
@@ -34,6 +42,25 @@ export function AdvancedSearchFilters({
         }),
         [t],
     );
+
+    const componentLabels = useMemo(() => {
+        return new Set(
+            componentsGroups.flatMap((componentGroup) => componentGroup.components).map((component) => component.label.toLowerCase()),
+        );
+    }, []);
+
+    const nodeTypes = useMemo(() => {
+        const availableTypes = allNodes
+            .flatMap((node) =>
+                selectorByName("type")
+                    .flatMap((s) => s.selector(node))
+                    .filter((item) => item !== undefined),
+            )
+            .map((selectorResult) => (typeof selectorResult === "string" ? selectorResult : selectorResult?.expression))
+            .filter((type) => componentLabels.has(type.toLowerCase()));
+
+        return uniq(availableTypes);
+    }, [allNodes]);
 
     useEffect(() => {
         const searchQuery = resolveSearchQuery(filter);
@@ -70,12 +97,7 @@ export function AdvancedSearchFilters({
             <SearchLabeledInput name="value" value={filterFields?.value || []} setFilterFields={setFilterFields}>
                 <SearchLabel label={displayNames["value"]} />
             </SearchLabeledInput>
-            <SearchLabeledAutocomplete
-                name="type"
-                options={useNodeTypes()}
-                value={filterFields?.type || []}
-                setFilterFields={setFilterFields}
-            >
+            <SearchLabeledAutocomplete name="type" options={nodeTypes} value={filterFields?.type || []} setFilterFields={setFilterFields}>
                 <SearchLabel label={displayNames["type"]} />
             </SearchLabeledAutocomplete>
             <SearchLabeledInput name="label" value={filterFields?.label || []} setFilterFields={setFilterFields}>
@@ -90,7 +112,16 @@ export function AdvancedSearchFilters({
             <SearchLabeledInput name="edge" value={filterFields?.edge || []} setFilterFields={setFilterFields}>
                 <SearchLabel label={displayNames["edge"]} />
             </SearchLabeledInput>
-            <Box sx={{ display: "flex", flexDirection: "row", justifyContent: "space-between", width: "100%", mt: 2, mb: 1 }}>
+            <Box
+                sx={{
+                    display: "flex",
+                    flexDirection: "row",
+                    justifyContent: "space-between",
+                    width: "100%",
+                    mt: 2,
+                    mb: 1,
+                }}
+            >
                 <Button sx={{ width: "45%" }} size="small" variant="outlined" onClick={handleClear}>
                     {t("search.panel.advancedFilters.clearButton.label", "Clear")}
                 </Button>
